LPS-54010 Move classes from web.impl package to just web, since impl is not really necessary.
