+++ modules/apps/item-selector/item-selector-web/src/com/liferay/item/selector/web/ItemSelectorRenderingImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.item.selector.web.impl;
+package com.liferay.item.selector.web;
 
 import com.liferay.item.selector.ItemSelectorRendering;
 import com.liferay.item.selector.ItemSelectorViewRenderer;
