+++ modules/apps/item-selector/item-selector-web/src/com/liferay/item/selector/web/ItemSelectorViewRendererImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.item.selector.web.impl;
+package com.liferay.item.selector.web;
 
 import com.liferay.item.selector.ItemSelectorCriterion;
 import com.liferay.item.selector.ItemSelectorView;
