+++ modules/apps/item-selector/item-selector-web/test/unit/com/liferay/item/selector/web/FlickrItemSelectorCriterionHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.item.selector.web.impl;
+package com.liferay.item.selector.web;
 
 import com.liferay.item.selector.ItemSelectorCriterionHandler;
 import com.liferay.item.selector.ItemSelectorView;
