+++ modules/apps/item-selector/item-selector-web/test/unit/com/liferay/item/selector/web/MediaItemSelectorCriterionHandler.java
@@ -1,0 +1,47 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.item.selector.web;
+
+import com.liferay.item.selector.ItemSelectorCriterionHandler;
+import com.liferay.item.selector.ItemSelectorView;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * @author Iván Zaera
+ */
+public class MediaItemSelectorCriterionHandler
+	implements ItemSelectorCriterionHandler<MediaItemSelectorCriterion> {
+
+	@Override
+	public Class<MediaItemSelectorCriterion> getItemSelectorCriterionClass() {
+		return MediaItemSelectorCriterion.class;
+	}
+
+	@Override
+	public List<ItemSelectorView<MediaItemSelectorCriterion>>
+		getItemSelectorViews(
+			MediaItemSelectorCriterion mediaItemSelectorCriterion) {
+
+		List<ItemSelectorView<MediaItemSelectorCriterion>> itemSelectorViews =
+			new ArrayList<>();
+
+		itemSelectorViews.add(new MediaItemSelectorView());
+
+		return itemSelectorViews;
+	}
+
+}
